requirements.txt
pandas==2.0.3	
sqlalchemy==1.4.50
apache-airflow[cncf.kubernetes]
pymysql
mysqlclient
tqdm